Schema registry sample upgrade - Iteration 1
Convert the vanilla applications that uses the Spring provided Scheam registry and Avro message converters.
Provide 2 sets of the same applications - one for the Kafka binder and another for the Rabbit binder.
The top level folder of each of these sets are named appropriately so that anyone who is interested in a
particular binder can easilty cd to that directory and run the samples